Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

NLPModelsJuMP.MathOptNLPModel

NLPModelsJuMP.MathOptNLSModel

NLPModelsJuMP.coo_sym_dot

NLPModelsJuMP.parser_JuMP

NLPModelsJuMP.parser_MOI

NLPModelsJuMP.parser_SAF

NLPModelsJuMP.parser_VAF

NLPModelsJuMP.parser_linear_expression

NLPModelsJuMP.parser_nonlinear_expression

NLPModelsJuMP.parser_objective_MOI

NLPModelsJuMP.replace!

​

NLPModelsJuMP.MathOptNLPModel — Method

MathOptNLPModel(model, hessian=true, name="Generic")

Construct a MathOptNLPModel from a JuMP model.

hessian should be set to false for multivariate user-defined functions registered without hessian.

NLPModelsJuMP.MathOptNLSModel — Method

MathOptNLSModel(model, F, name="Generic")

Construct a MathOptNLSModel from a JuMP model and a container of JuMP GenericAffExpr (generated by @expression) and NonlinearExpression (generated by @NLexpression).

NLPModelsJuMP.coo_sym_dot — Method

coo_sym_dot(rows, cols, vals, x, y)

Compute the product xᵀAy of a symmetric matrix A given by (rows, cols, vals) Only one triangle of A should be passed.

NLPModelsJuMP.parser_JuMP — Method

parser_JuMP(jmodel)

Parse variables informations of a JuMP.Model.

NLPModelsJuMP.parser_MOI — Method

parser_MOI(moimodel)

Parse linear constraints of a MOI.ModelLike.

NLPModelsJuMP.parser_SAF — Method

parser_SAF(fun, set, linrows, lincols, linvals, nlin, lin_lcon, lin_ucon)

Parse a ScalarAffineFunction fun with its associated set. linrows, lincols, linvals, lin_lcon and lin_ucon are updated.

NLPModelsJuMP.parser_VAF — Method

parser_VAF(fun, set, linrows, lincols, linvals, nlin, lin_lcon, lin_ucon)

Parse a VectorAffineFunction fun with its associated set. linrows, lincols, linvals, lin_lcon and lin_ucon are updated.

NLPModelsJuMP.parser_linear_expression — Method

parser_linear_expression(cmodel, nvar, F)

Parse linear expressions of type GenericAffExpr{Float64,VariableRef}.

NLPModelsJuMP.parser_nonlinear_expression — Method

parser_nonlinear_expression(cmodel, nvar, F)

Parse nonlinear expressions of type NonlinearExpression.

NLPModelsJuMP.parser_objective_MOI — Method

parser_objective_MOI(moimodel, nvar)

Parse linear and quadratic objective of a MOI.ModelLike.

NLPModelsJuMP.replace! — Method

replace!(ex, x)

Walk the expression ex and substitute in the actual variables x.