Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

NLPModelsJuMP.MathOptNLPModel

NLPModelsJuMP.MathOptNLSModel

NLPModelsJuMP.add_constraint_model

NLPModelsJuMP.coo_sym_add_mul!

NLPModelsJuMP.coo_sym_dot

NLPModelsJuMP.parser_MOI

NLPModelsJuMP.parser_NL

NLPModelsJuMP.parser_SAF

NLPModelsJuMP.parser_SQF

NLPModelsJuMP.parser_VAF

NLPModelsJuMP.parser_VQF

NLPModelsJuMP.parser_linear_expression

NLPModelsJuMP.parser_nonlinear_expression

NLPModelsJuMP.parser_objective_MOI

NLPModelsJuMP.parser_variables

NLPModelsJuMP.replace!

​

NLPModelsJuMP.MathOptNLPModel — Method

MathOptNLPModel(model, hessian=true, name="Generic")

Construct a MathOptNLPModel from a JuMP model.

hessian should be set to false for multivariate user-defined functions registered without hessian.

NLPModelsJuMP.MathOptNLSModel — Method

MathOptNLSModel(model, F, hessian=true, name="Generic")

Construct a MathOptNLSModel from a JuMP model and a container of JuMP GenericAffExpr (generated by @expression) and NonlinearExpression (generated by @NLexpression).

hessian should be set to false for multivariate user-defined functions registered without hessian.

NLPModelsJuMP.add_constraint_model — Method

add_constraint_model(Fmodel, Fi)

Add the nonlinear constraint Fi == 0 to the model Fmodel. If Fi is an Array, then we iterate over each component.

NLPModelsJuMP.coo_sym_add_mul! — Method

coo_sym_add_mul!(rows, cols, vals, x, y, α)

Update of the form y ← y + αAx where A is a symmetric matrix given by (rows, cols, vals). Only one triangle of A should be passed.

NLPModelsJuMP.coo_sym_dot — Method

coo_sym_dot(rows, cols, vals, x, y)

Compute the product xᵀAy of a symmetric matrix A given by (rows, cols, vals). Only one triangle of A should be passed.

NLPModelsJuMP.parser_MOI — Method

parser_MOI(moimodel, index_map, nvar)

Parse linear constraints of a MOI.ModelLike.

NLPModelsJuMP.parser_NL — Method

parser_NL(nlp_data; hessian)

Parse nonlinear constraints of an nlp_data.

NLPModelsJuMP.parser_SAF — Method

parser_SAF(fun, set, linrows, lincols, linvals, nlin, lin_lcon, lin_ucon, index_map)

Parse a ScalarAffineFunction fun with its associated set. linrows, lincols, linvals, lin_lcon and lin_ucon are updated.

NLPModelsJuMP.parser_SQF — Method

parser_SQF(fun, set, nvar, qcons, quad_lcon, quad_ucon, index_map)

Parse a ScalarQuadraticFunction fun with its associated set. qcons, quad_lcon, quad_ucon are updated.

NLPModelsJuMP.parser_VAF — Method

parser_VAF(fun, set, linrows, lincols, linvals, nlin, lin_lcon, lin_ucon, index_map)

Parse a VectorAffineFunction fun with its associated set. linrows, lincols, linvals, lin_lcon and lin_ucon are updated.

NLPModelsJuMP.parser_VQF — Method

parser_VQF(fun, set, nvar, qcons, quad_lcon, quad_ucon, index_map)

Parse a VectorQuadraticFunction fun with its associated set. qcons, quad_lcon, quad_ucon are updated.

NLPModelsJuMP.parser_linear_expression — Method

parser_linear_expression(cmodel, nvar, index_map, F)

Parse linear expressions of type VariableRef and GenericAffExpr{Float64,VariableRef}.

NLPModelsJuMP.parser_nonlinear_expression — Method

parser_nonlinear_expression(cmodel, nvar, F; hessian)

Parse nonlinear expressions of type NonlinearExpression.

NLPModelsJuMP.parser_objective_MOI — Method

parser_objective_MOI(moimodel, nvar, index_map)

Parse linear and quadratic objective of a MOI.ModelLike.

NLPModelsJuMP.parser_variables — Method

parser_variables(model)

Parse variables informations of a MOI.ModelLike.

NLPModelsJuMP.replace! — Method

replace!(ex, x)

Walk the expression ex and substitute in the actual variables x.