Tutorial

Tutorial

NLPModelsJuMP is a combination of NLPModels and JuMP, as the name implies. Sometimes it may be required to refer to the specific documentation, as we'll present here only the documention specific to NLPModelsJuMP.

Tutorial

MathProgNLPModel

MathProgNLSModel Tutorial

NLPtoMPB - Convert NLP to MathProgBase

MathProgNLPModel

NLPModelsJuMP.MathProgNLPModel — Type.

MathProgNLPModel(model, name="Generic")

Construct a MathProgNLPModel from a MathProgModel.

MathProgNLPModel(model; kwargs...)

Construct a MathProgNLPModel from a JuMP Model.

MathProgNLPModel is a simple yet efficient model. It uses JuMP to define the problem, and can be accessed through the NLPModels API. An advantage of MathProgNLPModel over simpler models such as ADNLPModels is that they provide sparse derivates.

Let's define the famous Rosenbrock function

\[f(x) = (x_1 - 1)^2 + 100(x_2 - x_1^2)^2,\]

with starting point $x^0 = (-1.2,1.0)$.

using NLPModels, NLPModelsJuMP, JuMP

x0 = [-1.2; 1.0]
model = Model() # No solver is required
@variable(model, x[i=1:2], start=x0[i])
@NLobjective(model, Min, (x[1] - 1)^2 + 100 * (x[2] - x[1]^2)^2)

nlp = MathProgNLPModel(model)

MathProgNLPModel(Minimization problem Generic
nvar = 2, ncon = 0 (0 linear)
, NLPModelsJuMP.MathProgModel(JuMP.NLPEvaluator(Minimization problem with:
 * 0 linear constraints
 * 2 variables
Solver is NLPModelsJuMP.ModelReader(), 0×2 SparseArrays.SparseMatrixCSC{Float64,Int64} with 0 stored entries, Float64[], true, [0.0, 0.0], JuMP.FunctionStorage(ReverseDiffSparse.NodeData[NodeData(CALL, 1, -1), NodeData(CALL, 4, 1), NodeData(CALL, 2, 2), NodeData(VARIABLE, 1, 3), NodeData(VALUE, 1, 3), NodeData(VALUE, 2, 2), NodeData(CALL, 3, 1), NodeData(VALUE, 3, 7), NodeData(CALL, 4, 7), NodeData(CALL, 2, 9), NodeData(VARIABLE, 2, 10), NodeData(CALL, 4, 10), NodeData(VARIABLE, 1, 12), NodeData(VALUE, 4, 12), NodeData(VALUE, 5, 9)],
  [2 ,  1]  =  true
  [7 ,  1]  =  true
  [3 ,  2]  =  true
  [6 ,  2]  =  true
  [4 ,  3]  =  true
  [5 ,  3]  =  true
  [8 ,  7]  =  true
  [9 ,  7]  =  true
  [10,  9]  =  true
  [15,  9]  =  true
  [11, 10]  =  true
  [12, 10]  =  true
  [13, 12]  =  true
  [14, 12]  =  true, [1.0, 2.0, 100.0, 2.0, 2.0], [24.2, 4.84, -2.2, -1.2, 1.0, 2.0, 19.36, 100.0, 0.1936, -0.44, 1.0, 1.44, -1.2, 2.0, 2.0], [0.0, 1.0, -4.4, 1.0, -1.0, NaN, 1.0, 0.1936, 100.0, -0.88, 1.0, -1.0, -2.4, NaN, NaN], [1.0, 1.0, -4.4, -4.4, 0.0, 0.0, 1.0, 0.0, 100.0, -88.0, -88.0, 88.0, -211.2, 0.0, 0.0], [1, 2], [1, 2, 2], [1, 2, 1], ReverseDiffSparse.Coloring.RecoveryInfo(Array{Int64,1}[[2, 1]], Array{Int64,1}[[2, 1]], Array{Int64,1}[[0, 1]], [1, 2], 2, 1, [1, 2]), [1330.0 480.0; 480.0 200.0], NONLINEAR::Linearity = 3, Int64[]), JuMP.FunctionStorage[], JuMP.SubexpressionStorage[], Int64[], Float64[], Float64[], ReverseDiffSparse.Linearity[], Any[], [-1.2, 1.0], [6.93297e-310, 0.0], Float64[], false, true, [-215.6, -88.0, -4.4, 0.0, 1.0, 0.0, 1.0, 0.0, 0.0, 0.0  …  0.0, 1.0, -2.4, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 0.0], [0.0, 0.0, 0.0, 0.0, 2.0, 0.0, 0.0, 0.0, 0.0, 0.0  …  0.0, 0.0, 0.0, 0.0, 2.0, 0.0, NaN, NaN, NaN, NaN], [0.0, 0.0, 0.0, 0.0, 2.0, 0.0, 2.0, 0.0, 6.93295e-310, 6.93295e-310  …  480.0, 200.0, -480.0, -200.0, 1328.0, 480.0, 3.90312e-322, 4.05134e-322, 4.19956e-322, 4.34778e-322], [0.0, 0.0, 0.0, 0.0], [480.0, 0.0, 480.0, 200.0], Float64[], Float64[], [1, 2, 2], [1, 2, 1], 2, false, false, false, false, false), 2, 0, [-1.2, 1.0], Float64[], [-Inf, -Inf], [Inf, Inf], Float64[], Float64[], :Min, :Uninitialized), NLPModels.Counters(0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0), Int64[], Int64[], Float64[], [1, 2, 2], [1, 2, 1], [0.0, 0.0, 0.0])

Let's get the objective function value at $x^0$, using only nlp.

fx = obj(nlp, nlp.meta.x0)
println("fx = $fx")

fx = 24.199999999999996

Let's try the gradient and Hessian.

gx = grad(nlp, nlp.meta.x0)
Hx = hess(nlp, nlp.meta.x0)
println("gx = $gx")
println("Hx = $Hx")

gx = [-215.6, -88.0]
Hx =
  [1, 1]  =  1330.0
  [2, 1]  =  480.0
  [2, 2]  =  200.0

Notice how only the lower triangle of the Hessian is stored, which is the default for NLPModels.

Let's do something a little more complex here, defining a function to try to solve this problem through steepest descent method with Armijo search. Namely, the method

Given $x^0$, $\varepsilon > 0$, and $\eta \in (0,1)$. Set $k = 0$;

If $\Vert \nabla f(x^k) \Vert < \varepsilon$ STOP with $x^* = x^k$;

Compute $d^k = -\nabla f(x^k)$;

Compute $\alpha_k \in (0,1]$ such that $f(x^k + \alpha_kd^k) < f(x^k) + \alpha_k\eta \nabla f(x^k)^Td^k$

Define $x^{k+1} = x^k + \alpha_kx^k$

Update $k = k + 1$ and go to step 2.

using LinearAlgebra

function steepest(nlp; itmax=100000, eta=1e-4, eps=1e-6, sigma=0.66)
  x = nlp.meta.x0
  fx = obj(nlp, x)
  ∇fx = grad(nlp, x)
  slope = dot(∇fx, ∇fx)
  ∇f_norm = sqrt(slope)
  iter = 0
  while ∇f_norm > eps && iter < itmax
    t = 1.0
    x_trial = x - t * ∇fx
    f_trial = obj(nlp, x_trial)
    while f_trial > fx - eta * t * slope
      t *= sigma
      x_trial = x - t * ∇fx
      f_trial = obj(nlp, x_trial)
    end
    x = x_trial
    fx = f_trial
    ∇fx = grad(nlp, x)
    slope = dot(∇fx, ∇fx)
    ∇f_norm = sqrt(slope)
    iter += 1
  end
  optimal = ∇f_norm <= eps
  return x, fx, ∇f_norm, optimal, iter
end

x, fx, ngx, optimal, iter = steepest(nlp)
println("x = $x")
println("fx = $fx")
println("ngx = $ngx")
println("optimal = $optimal")
println("iter = $iter")

x = [1.0, 1.0]
fx = 4.2438440239813445e-13
ngx = 9.984661274466944e-7
optimal = true
iter = 17962

Maybe this code is too complicated? If you're in a class you just want to show a Newton step.

f(x) = obj(nlp, x)
g(x) = grad(nlp, x)
H(x) = Symmetric(hess(nlp, x), :L)
x = nlp.meta.x0
d = -H(x)\g(x)

2-element Array{Float64,1}:
 0.02471910112359557
 0.3806741573033706

or a few

for i = 1:5
  global x
  x = x - H(x)\g(x)
  println("x = $x")
end

x = [-1.17528, 1.38067]
x = [0.763115, -3.17503]
x = [0.76343, 0.582825]
x = [0.999995, 0.944027]
x = [0.999996, 0.999991]

Notice how we can use the method with different NLPModels:

f(x) = (x[1] - 1.0)^2 + 100 * (x[2] - 1.0)^2

adnlp = ADNLPModel(f, x0)
x, fx, ngx, optimal, iter = steepest(adnlp)

([1.0, 1.0], 6.745975419202182e-14, 5.194603129865527e-7, true, 14)

OptimizationProblems

The package OptimizationProblems provides a collection of problems defined in JuMP format, which can be converted to MathProgNLPModel.

using OptimizationProblems # Defines a lot of JuMP models

nlp = MathProgNLPModel(woods())
x, fx, ngx, optimal, iter = steepest(nlp)
println("fx = $fx")
println("ngx = $ngx")
println("optimal = $optimal")
println("iter = $iter")

fx = 1.0000000000002167
ngx = 9.893253859340887e-7
optimal = true
iter = 12016

Constrained problems can also be converted.

using NLPModels, NLPModelsJuMP, JuMP

model = Model()
x0 = [-1.2; 1.0]
@variable(model, x[i=1:2] >= 0.0, start=x0[i])
@NLobjective(model, Min, (x[1] - 1)^2 + 100 * (x[2] - x[1]^2)^2)
@constraint(model, x[1] + x[2] == 3.0)
@NLconstraint(model, x[1] * x[2] >= 1.0)

nlp = MathProgNLPModel(model)

println("cx = $(cons(nlp, nlp.meta.x0))")
println("Jx = $(jac(nlp, nlp.meta.x0))")

cx = [-0.2, -2.2]
Jx =
  [1, 1]  =  1.0
  [2, 1]  =  1.0
  [1, 2]  =  1.0
  [2, 2]  =  -1.2

MathProgNLSModel Tutorial

NLPModelsJuMP.MathProgNLSModel — Type.

Construct a MathProgNLSModel from two MathProgModels.

MathProgNLSModel(cmodel, F)

Construct a MathProgNLSModel from a JuMP Model and a vector of NLexpression.

MathProgNLSModel is a model for nonlinear least squares using JuMP, The objective function of NLS problems has the form $f(x) = \tfrac{1}{2}\|F(x)\|^2$, but specialized methods handle $F$ directly, instead of $f$. To use MathProgNLSModel, we define a JuMP model without the objective, and use NLexpressions to define the residual function $F$. For instance, the Rosenbrock function can be expressed in nonlinear least squares format by defining

\[F(x) = \begin{bmatrix} x_1 - 1\\ 10(x_2 - x_1^2) \end{bmatrix},\]

and noting that $f(x) = \|F(x)\|^2$ (the constant $\frac{1}{2}$ is ignored as it doesn't change the solution). We implement this function as

using NLPModels, NLPModelsJuMP, JuMP

model = Model()
x0 = [-1.2; 1.0]
@variable(model, x[i=1:2], start=x0[i])
@NLexpression(model, F1, x[1] - 1)
@NLexpression(model, F2, 10 * (x[2] - x[1]^2))

nls = MathProgNLSModel(model, [F1, F2], name="rosen-nls")

residual(nls, nls.meta.x0)

2-element Array{Float64,1}:
 -2.2
 -4.3999999999999995

jac_residual(nls, nls.meta.x0)

2×2 SparseArrays.SparseMatrixCSC{Float64,Int64} with 3 stored entries:
  [1, 1]  =  1.0
  [2, 1]  =  24.0
  [2, 2]  =  10.0

NLPtoMPB - Convert NLP to MathProgBase

NLPModelsJuMP.NLPtoMPB — Function.

mp = NLPtoMPB(nlp, solver)

Return a MathProgBase model corresponding to an AbstractNLPModel.

Arguments

nlp::AbstractNLPModel

solver::AbstractMathProgSolver a solver instance, e.g., IpoptSolver()

Currently, all models are treated as nonlinear models.

Return values

The function returns a MathProgBase model mpbmodel such that it should be possible to call

MathProgBase.optimize!(mpbmodel)

In addition to creating NLPModels using JuMP, we might want to convert an NLPModel to a MathProgBase model to use the solvers available. For instance

using Ipopt, NLPModels, NLPModelsJuMP, LinearAlgebra, JuMP, MathProgBase

nlp = ADNLPModel(x -> dot(x, x), ones(2),
                 c=x->[x[1] + 2 * x[2] - 1.0], lcon=[0.0], ucon=[0.0])
model = NLPtoMPB(nlp, IpoptSolver())

MathProgBase.optimize!(model)

******************************************************************************
This program contains Ipopt, a library for large-scale nonlinear optimization.
 Ipopt is released as open source code under the Eclipse Public License (EPL).
         For more information visit http://projects.coin-or.org/Ipopt
******************************************************************************

This is Ipopt version 3.12.8, running with linear solver mumps.
NOTE: Other linear solvers might be more efficient (see Ipopt documentation).

Number of nonzeros in equality constraint Jacobian...:        2
Number of nonzeros in inequality constraint Jacobian.:        0
Number of nonzeros in Lagrangian Hessian.............:        3

Total number of variables............................:        2
                     variables with only lower bounds:        0
                variables with lower and upper bounds:        0
                     variables with only upper bounds:        0
Total number of equality constraints.................:        1
Total number of inequality constraints...............:        0
        inequality constraints with only lower bounds:        0
   inequality constraints with lower and upper bounds:        0
        inequality constraints with only upper bounds:        0

iter    objective    inf_pr   inf_du lg(mu)  ||d||  lg(rg) alpha_du alpha_pr  ls
   0  2.0000000e+00 2.00e+00 8.00e-01  -1.0 0.00e+00    -  0.00e+00 0.00e+00   0
   1  2.0000000e-01 0.00e+00 1.11e-16  -1.0 8.00e-01    -  1.00e+00 1.00e+00f  1

Number of Iterations....: 1

                                   (scaled)                 (unscaled)
Objective...............:   2.0000000000000001e-01    2.0000000000000001e-01
Dual infeasibility......:   1.1102230246251565e-16    1.1102230246251565e-16
Constraint violation....:   0.0000000000000000e+00    0.0000000000000000e+00
Complementarity.........:   0.0000000000000000e+00    0.0000000000000000e+00
Overall NLP error.......:   1.1102230246251565e-16    1.1102230246251565e-16

Number of objective function evaluations             = 2
Number of objective gradient evaluations             = 2
Number of equality constraint evaluations            = 2
Number of inequality constraint evaluations          = 0
Number of equality constraint Jacobian evaluations   = 2
Number of inequality constraint Jacobian evaluations = 0
Number of Lagrangian Hessian evaluations             = 1
Total CPU secs in IPOPT (w/o function evaluations)   =      0.196
Total CPU secs in NLP function evaluations           =      0.036

EXIT: Optimal Solution Found.
0